Fix sequence_number type

+++ src/media/media.rs
@@ -113,8 +113,10 @@ pub struct Episode {
     pub episode: String,
     #[serde(deserialize_with = "crate::internal::serde::deserialize_maybe_null_to_default")]
     pub episode_number: u32,
-    /// Usually the same as [`Episode::episode_number`], I don't know the purpose of this.
-    pub sequence_number: u32,
+    /// Usually also the same as [`Episode::episode_number`]. If the episode number is null (which
+    /// occurs for the first AOT episode, which is a preview, for example) this might be a floating
+    /// number like 0.5.
+    pub sequence_number: f32,
     #[serde(alias = "duration_ms")]
     #[serde(deserialize_with = "crate::internal::serde::deserialize_millis_to_duration")]
     #[default(Duration::milliseconds(0))]
